fastapi-server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
requests==2.31.0
python-dotenv==1.0.0
